tasker2/requirements.txt
django
django-filter
django-crispy-forms
